refactor: Moved content into App
